Defines protocol between gdbserver and a 2LS.
The protocol is called D9.

This will be part of parlib and in its current form can be used with any
2LS.

The protocol is intended to be synchronous on the 2LS side and asynchronous
on the gdbserver side. It looks a lot like 9P (hence, the name). Some of
the issues around async are not solved yet; e.g. the client side
(gdbserver) currently expects all operations to be serialized without
interleaving responses. E.g. it could not currently handle this sequence of
operations:

TWAIT
TREADMEM
RWAIT
RREADMEM

Change-Id: Ib7c8fd800b1550f6b72ce0feae57c7f5dbdf11f6
Signed-off-by: Christopher Koch <chrisko@google.com>