Remove our glibc poll implementation (XCC)
The old one just immediately returned.  Now that we have a version of
poll() in iplib, that one would have overridden glibc's.  However, if we
messed up and didn't link with iplib, then we'd silently be using the old
broken glibc version again.  This way, we'll catch it with a stub warning.

Rebuild glibc.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>