x86: Secure eflags when securing contexts
We had been making sure interrupts was on, but we weren't enforcing some of
the other flags.  I don't know if this is all of them, but it's the list of
ones that sounded dangerous from the SDM.

The IOPL one was probably a security hole.  A process could have asked the
kernel to pop a HWTF that had IOPL set to 3, and then started mucking with
devices.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>