Reduce mmap calls in vfprintf.c (XCC)
It turns out that the work_buffer path gets called a lot, including by
bash.  I saw a lot of mmaps in a few traces.  I think we can put that 1k on
the stack, since it's right near the magic number of 1/4 of a minimum
stack size.  We'll see.

This also fixes the other mmap, which should have been asking for
MAP_ANONYMOUS.

Rebuild glibc.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>