Fix offset calculation in populate_va
That should clearly be a +, not a -, since we're figuring out how far into
the VMR to map.

This would trigger if you had a file mmapped that wasn't MAP_POPULATE, then
had a uthread fault on accessing that file, and it wasn't on the first page
of the file.

While we're here, we can also catch any integer overflows with offset and
length.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>